Added strava info accountform and userform
