+++ client/src/App.js
@@ -14,7 +14,7 @@ import Activities from './components/Activity/Activities';
 import ActivityPage from './components/Activity/ActivityPage';
 import ActivityForm from './components/Activity/ActivityForm';
 import UserPage from './components/User/UserPage';
-import UserForm from './components/User/UserForm';
+import UserForm from './components/User/UserForm.jsx';
 import Register from './components/Auth/Register/Register';
 import Challenges from './components/Challenges/Challenges';
 import Teams from './components/Team/Teams';
