+++ client/src/components/User/UserPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import UserForm from "./UserForm";
+import UserForm from "./UserForm.jsx";
   
 class UserPage extends React.Component {
